Added Navigation structure with boilerplate examples for 2 different flows
Though this structure may change later, for now I have broken up the folder structure for the screens in to separate folders for each flow of screens.  Each flow directory has a flow tsx file with the navigation setup for that flow.  For type sensitivity, there is a Navigation.tsx file that handles all the strings, screen parameters, etc. that can then be referenced by any other component.