Added Common package with master config setup for client & server packages (#1)
Added a common package with a config setup that allows devs to add systems with nested systems and parameters that the web client, mobile client, and node server have access to.  Also used nodemon and concurrently to allow devs to run both the server and either client at the same time, while also using nodemon to listen for any file changes to the config, auto generating a new config object with every single system being nested right in the root of the object, allowing for quick access to any system or paramter.  Will eventually look at moving away from nodemon for this process and using some sort of file wathcing package instead, was just having issues with closures when using node-watch or fb-watchman